Fix OnIndexSeries ref type
